forsamples/requirements.txt
requests
xmltodict
python-dotenv
mysql-connector-python
python-dotenv